fix issue with empty PID

+++ dot_config/fish/conf.d/executable_update_git_status.fish
@@ -2,6 +2,10 @@
 
 function update_git_info
     set pid $argv[1]
+    # fix issue?
+    if test "$pid" = ""
+        exit
+    end
 
     # XXX: global variables takes time to update so we need to wait 1ms
     trap "kill -s SIGUSR1 $pid" EXIT
